Added methods isLockedByCurrentThread() and isLockedBy(Thread thread).
1) Added method isLockedByCurrentThread() to the IMap interface. We
need this in our safe-unlock check – we first do unlock, then we check
if the lock has really been removed (the unlock method is a nasty one,
it suppresses the timeout exception and returns as if nothing has
happened). But our check needs to be more sophisticated than that – if
another thread picks up the lock immediately, we still get information
that it is locked and enter the "force-unlock" sequence. Therefore, the
check isLocked -> isLockedByCurrentThread.
2) Added another method to both lock and map APIs –
isLockedBy(Thread thread). My idea is to use it in the "force-unlock"
part of our code – there, before force unlocking, we should check if the
thread which couldn’t release the lock is still the owner – if it isn’t,
we return as if we have managed to unlock. In this manner, we guarantee
that the force-unlock won’t cause someone who properly acquired it to
release it too early.